Move ModelMakerCallback to interfaces
